+++ crates/miden-testing/src/kernel_tests/tx/test_asset_vault.rs
@@ -35,7 +35,7 @@ fn test_get_balance() -> anyhow::Result<()> {
     let faucet_id: AccountId = ACCOUNT_ID_PUBLIC_FUNGIBLE_FAUCET.try_into().unwrap();
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::account
 
         begin
@@ -71,7 +71,7 @@ fn test_get_balance_non_fungible_fails() -> anyhow::Result<()> {
     let faucet_id = AccountId::try_from(ACCOUNT_ID_PUBLIC_NON_FUNGIBLE_FAUCET).unwrap();
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::account
 
         begin
@@ -105,7 +105,7 @@ fn test_has_non_fungible_asset() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::account
 
         begin
@@ -146,7 +146,7 @@ fn test_add_fungible_asset_success() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -196,7 +196,7 @@ fn test_add_non_fungible_asset_fail_overflow() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -230,7 +230,7 @@ fn test_add_non_fungible_asset_success() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -275,7 +275,7 @@ fn test_add_non_fungible_asset_fail_duplicate() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -315,7 +315,7 @@ fn test_remove_fungible_asset_success_no_balance_remaining() -> anyhow::Result<(
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -363,7 +363,7 @@ fn test_remove_fungible_asset_fail_remove_too_much() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -402,7 +402,7 @@ fn test_remove_fungible_asset_success_balance_remaining() -> anyhow::Result<()> 
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -454,7 +454,7 @@ fn test_remove_inexisting_non_fungible_asset_fails() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin
@@ -493,7 +493,7 @@ fn test_remove_non_fungible_asset_success() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
 
         begin