+++ crates/miden-testing/src/kernel_tests/tx/test_epilogue.rs
@@ -69,9 +69,9 @@ fn test_epilogue() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
-        use.kernel::account
-        use.kernel::epilogue
+        use.$kernel::prologue
+        use.$kernel::account
+        use.$kernel::epilogue
 
         {output_notes_data_procedure}
 
@@ -171,8 +171,8 @@ fn test_compute_output_note_id() -> anyhow::Result<()> {
     for (note, i) in tx_context.expected_output_notes().iter().zip(0u32..) {
         let code = format!(
             "
-            use.kernel::prologue
-            use.kernel::epilogue
+            use.$kernel::prologue
+            use.$kernel::epilogue
 
             {output_notes_data_procedure}
 
@@ -258,9 +258,9 @@ fn test_epilogue_asset_preservation_violation_too_few_input() -> anyhow::Result<
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
-        use.kernel::epilogue
+        use.$kernel::epilogue
 
         {output_notes_data_procedure}
 
@@ -341,9 +341,9 @@ fn test_epilogue_asset_preservation_violation_too_many_fungible_input() -> anyho
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
-        use.kernel::epilogue
+        use.$kernel::epilogue
 
         {output_notes_data_procedure}
 
@@ -373,10 +373,10 @@ fn test_block_expiration_height_monotonically_decreases() -> anyhow::Result<()> 
 
     let test_pairs: [(u64, u64); 3] = [(9, 12), (18, 3), (20, 20)];
     let code_template = "
-        use.kernel::prologue
-        use.kernel::tx
-        use.kernel::epilogue
-        use.kernel::account
+        use.$kernel::prologue
+        use.$kernel::tx
+        use.$kernel::epilogue
+        use.$kernel::account
 
         begin
             exec.prologue::prepare_transaction
@@ -421,7 +421,7 @@ fn test_invalid_expiration_deltas() -> anyhow::Result<()> {
 
     let test_values = [0u64, u16::MAX as u64 + 1, u32::MAX as u64];
     let code_template = "
-        use.kernel::tx
+        use.$kernel::tx
 
         begin
             push.{value_1}
@@ -447,10 +447,10 @@ fn test_no_expiration_delta_set() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
 
     let code_template = "
-    use.kernel::prologue
-    use.kernel::epilogue
-    use.kernel::tx
-    use.kernel::account
+    use.$kernel::prologue
+    use.$kernel::epilogue
+    use.$kernel::tx
+    use.$kernel::account
 
     begin
         exec.prologue::prepare_transaction
@@ -483,10 +483,10 @@ fn test_epilogue_increment_nonce_success() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
-        use.kernel::epilogue
-        use.kernel::memory
+        use.$kernel::epilogue
+        use.$kernel::memory
 
         begin
             exec.prologue::prepare_transaction
@@ -531,9 +531,9 @@ fn test_epilogue_increment_nonce_violation() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.test::account
-        use.kernel::epilogue
+        use.$kernel::epilogue
 
         {output_notes_data_procedure}
 
@@ -589,9 +589,9 @@ fn test_epilogue_empty_transaction_with_empty_output_note() -> anyhow::Result<()
     let tx_script_source = format!(
         r#"
         use.miden::tx
-        use.kernel::prologue
-        use.kernel::epilogue
-        use.kernel::note
+        use.$kernel::prologue
+        use.$kernel::epilogue
+        use.$kernel::note
 
         begin
             exec.prologue::prepare_transaction