+++ crates/miden-testing/src/kernel_tests/tx/test_fpi.rs
@@ -110,7 +110,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
         "
         use.std::sys
         
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::tx
 
         begin
@@ -160,7 +160,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
         "
         use.std::sys
 
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::tx
 
         begin
@@ -216,7 +216,7 @@ fn test_fpi_memory() -> anyhow::Result<()> {
         "
         use.std::sys
 
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::tx
 
         begin
@@ -377,7 +377,7 @@ fn test_fpi_memory_two_accounts() -> anyhow::Result<()> {
         "
         use.std::sys
 
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::tx
 
         begin
@@ -1232,7 +1232,7 @@ fn test_fpi_stale_account() -> anyhow::Result<()> {
         "
       use.std::sys
 
-      use.kernel::prologue
+      use.$kernel::prologue
       use.miden::tx
 
       begin