+++ crates/miden-testing/src/kernel_tests/tx/test_note.rs
@@ -47,8 +47,8 @@ fn test_get_sender_no_sender() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
     // calling get_sender should return sender
     let code = "
-        use.kernel::memory
-        use.kernel::prologue
+        use.$kernel::memory
+        use.$kernel::prologue
         use.miden::note
 
         begin
@@ -86,8 +86,8 @@ fn test_get_sender() -> anyhow::Result<()> {
 
     // calling get_sender should return sender
     let code = "
-        use.kernel::prologue
-        use.kernel::note->note_internal
+        use.$kernel::prologue
+        use.$kernel::note->note_internal
         use.miden::note
 
         begin
@@ -144,8 +144,8 @@ fn test_get_vault_data() -> anyhow::Result<()> {
         "
         use.std::sys
 
-        use.kernel::prologue
-        use.kernel::note
+        use.$kernel::prologue
+        use.$kernel::note
 
         begin
             exec.prologue::prepare_transaction
@@ -235,8 +235,8 @@ fn test_get_assets() -> anyhow::Result<()> {
         "
         use.std::sys
 
-        use.kernel::prologue
-        use.kernel::note->note_internal
+        use.$kernel::prologue
+        use.$kernel::note->note_internal
         use.miden::note
 
         proc.process_note_0
@@ -358,8 +358,8 @@ fn test_get_inputs() -> anyhow::Result<()> {
 
     let code = format!(
         "
-        use.kernel::prologue
-        use.kernel::note->note_internal
+        use.$kernel::prologue
+        use.$kernel::note->note_internal
         use.miden::note
 
         begin
@@ -456,7 +456,7 @@ fn test_get_exactly_8_inputs() -> anyhow::Result<()> {
         .build()?;
 
     let tx_code = "
-            use.kernel::prologue
+            use.$kernel::prologue
             use.miden::note
 
             begin
@@ -498,8 +498,8 @@ fn test_note_setup() -> anyhow::Result<()> {
     };
 
     let code = "
-        use.kernel::prologue
-        use.kernel::note
+        use.$kernel::prologue
+        use.$kernel::note
 
         begin
             exec.prologue::prepare_transaction
@@ -553,9 +553,9 @@ fn test_note_script_and_note_args() -> miette::Result<()> {
     };
 
     let code = "
-        use.kernel::prologue
-        use.kernel::memory
-        use.kernel::note
+        use.$kernel::prologue
+        use.$kernel::memory
+        use.$kernel::note
 
         begin
             exec.prologue::prepare_transaction
@@ -630,7 +630,7 @@ fn test_get_note_serial_number() -> anyhow::Result<()> {
 
     // calling get_serial_number should return the serial number of the note
     let code = "
-        use.kernel::prologue
+        use.$kernel::prologue
         use.miden::note
 
         begin
@@ -766,7 +766,7 @@ fn test_get_current_script_root() -> anyhow::Result<()> {
 
     // calling get_script_root should return script root
     let code = "
-    use.kernel::prologue
+    use.$kernel::prologue
     use.miden::note
 
     begin
@@ -816,8 +816,8 @@ fn test_build_note_metadata() -> miette::Result<()> {
     for (iteration, test_metadata) in [test_metadata1, test_metadata2].into_iter().enumerate() {
         let code = format!(
             "
-        use.kernel::prologue
-        use.kernel::tx
+        use.$kernel::prologue
+        use.$kernel::tx
 
         begin
           exec.prologue::prepare_transaction