+++ crates/miden-testing/src/kernel_tests/tx/test_prologue.rs
@@ -78,7 +78,7 @@ fn test_transaction_prologue() -> anyhow::Result<()> {
     };
 
     let code = "
-        use.kernel::prologue
+        use.$kernel::prologue
 
         begin
             exec.prologue::prepare_transaction
@@ -500,7 +500,7 @@ pub fn create_account_test(
         .unwrap();
 
     let code = "
-  use.kernel::prologue
+  use.$kernel::prologue
 
   begin
       exec.prologue::prepare_transaction
@@ -665,7 +665,7 @@ pub fn create_account_invalid_seed() -> anyhow::Result<()> {
         .build()?;
 
     let code = "
-      use.kernel::prologue
+      use.$kernel::prologue
 
       begin
           exec.prologue::prepare_transaction
@@ -683,8 +683,8 @@ pub fn create_account_invalid_seed() -> anyhow::Result<()> {
 fn test_get_blk_version() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
     let code = "
-    use.kernel::memory
-    use.kernel::prologue
+    use.$kernel::memory
+    use.$kernel::prologue
 
     begin
         exec.prologue::prepare_transaction
@@ -706,8 +706,8 @@ fn test_get_blk_version() -> anyhow::Result<()> {
 fn test_get_blk_timestamp() -> anyhow::Result<()> {
     let tx_context = TransactionContextBuilder::with_existing_mock_account().build()?;
     let code = "
-    use.kernel::memory
-    use.kernel::prologue
+    use.$kernel::memory
+    use.$kernel::prologue
 
     begin
         exec.prologue::prepare_transaction